+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add jdbi3 module (#2)
+
 ### 0.4.0 (2018-06-25)
 * Update to guicey 4.2.0 (dropwizard 1.3.5)
 * Fix jdbi compatibility with guice 4.2.0